requirements.txt
streamlit
openai
langchain
weaviate-client
tiktoken
langchain-openai